11 - Capstone project/requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
python-dotenv
